% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalpp.R
\name{addnull}
\alias{addnull}
\title{addnull}
\usage{
addnull(what, val)
}
\arguments{
\item{what}{vector or matrix}

\item{val}{value to add for the null model}
}
\value{
what with an additional first entry or first row filled with val
}
\description{
Internal function to add a null entry
}
\author{
Chris Wallace
}
