% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTFM-functions.R
\name{BFexpanded.fn}
\alias{BFexpanded.fn}
\title{Expand tag models and calculate joint BFs at expanded models}
\usage{
BFexpanded.fn(ppbf, tags, t1snp.data, t1pheno, t2snp.data, t2pheno, mydir,
  trait1, trait2)
}
\arguments{
\item{ppbf}{output from PPBF.tags.fn}

\item{tags}{set of tag SNPs from common controla}

\item{t1snp.data}{a SnpMatrix object for trait 1}

\item{t1pheno}{a phenotype vector for trait 1}

\item{t2snp.data}{a SnpMatrix object for trait 2}

\item{t2pheno}{a phenotype vector for trait 2}

\item{trait1}{name of trait 1, for filename purposes}

\item{trait2}{name of trait 2, for filename purposes}
}
\author{
Jenn Asimit
expand tags at all marginal models that are part of a "best" joint model
at expanded tags, re-fit joint models 
merge expanded re-fit joint with expanded non-re-fit joint
}
