% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa.R
\name{calckappa}
\alias{calckappa}
\title{calckappa}
\usage{
calckappa(nsnps, p, target.odds)
}
\arguments{
\item{nsnps}{number of snps in the region}

\item{p}{probability any snp is causal (assumes a binomial distribution for the number of causal snps, with expected number = p * nsnps)}

\item{target.odds}{prior odds of any sharing of causal variants between any pair of traits}
}
\value{
the value of kappa corresponding to supplied target.odds
}
\description{
Calculate kappa for a given region and priors on the number of causal variants in the region, and the prior odds of causal variant sharing
}
\author{
Chris Wallace
}
