% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalpp.R
\name{calcpp}
\alias{calcpp}
\title{calcpp}
\usage{
calcpp(pr, lBF)
}
\arguments{
\item{pr}{vector or matrix of priors}

\item{lBF}{vector of log ABF}
}
\value{
vector or matrix of pp
}
\description{
Internal function to combine vector or matrix prior with vector of log ABF to get pp per model
}
\author{
Chris Wallace
}
