% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{check.fn}
\alias{check.fn}
\title{internal function for MPP.fn}
\usage{
check.fn(k, msep, out, gnames)
}
\arguments{
\item{k}{index}

\item{msep}{output from sep.fn}

\item{out}{storage vector for output}

\item{gnames}{SNP group names}
}
\description{
internal function for MPP.fn
}
\author{
Jenn Asimit
}
