% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTFM-functions.R
\name{expand.tags.bf}
\alias{expand.tags.bf}
\title{Expand tag SNP models}
\usage{
expand.tags.bf(best, tags)
}
\arguments{
\item{best}{is the output from T1mods.fn, a data.frame of the models and their sizes}

\item{tags}{is the set of tag SNPs; a tags object}
}
\value{
a list with three components mod (tag SNP model), size (model size), str (expanded tag SNP model)
}
