% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa.R
\name{kappa.opt.fn}
\alias{kappa.opt.fn}
\title{Kappa selection}
\usage{
\method{kappa}{opt.fn}(Nd, Ps2, nsnps)
}
\arguments{
\item{Nd}{Number of diseases to fine-map}

\item{Ps2}{Probability of any shared causal variant between a disease pair}

\item{n}{Number of SNPs in region}
}
\value{
Optimal kappa value
}
