% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa.R
\name{kappaOpt}
\alias{kappaOpt}
\title{Kappa selection}
\usage{
kappaOpt(Nd, Ps2, nsnps)
}
\arguments{
\item{Nd}{Number of diseases to fine-map}

\item{Ps2}{Probability of any shared causal variant between a disease pair}

\item{nsnps}{Number of SNPs in region}
}
\value{
Optimal kappa value
}
\description{
Kappa selection
}
