% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{make.snp.groups.fn}
\alias{make.snp.groups.fn}
\title{SNP groups that have marginal PP of inclusion (MPPi) above a threshold for some disease}
\usage{
make.snp.groups.fn(groups, rdis, mppi.thr = 0.05)
}
\arguments{
\item{groups}{List consisting of groups: SNP groups formed for a set of diseases; summary: for each group the min r2, max r2 and MPPi for each disease is given; r2: matrix of r2 between each pair of SNPs}

\item{rdis}{Full set of diseases for fine-mapping}

\item{mppi.thr}{Threshold for retaining SNP groups that have MPPi > mppi.thr for some disease in rdis}
}
\value{
List of SNP groups that summarise fine-mapping results
}
\description{
SNP groups that have marginal PP of inclusion (MPPi) above a threshold for some disease
}
