% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTFM-functions.R
\name{make.tags.fn}
\alias{make.tags.fn}
\title{Generate set of tag SNPs with LD r2}
\usage{
make.tags.fn(r2, G0, mydir = NA)
}
\arguments{
\item{r2}{r-squared for tag SNP generation
param G0 SnpMatrix object for control data
param mydir file to save tags object to, if specified}
}
\description{
Generate set of tag SNPs from control genotype matrix G0 ("SnpMatrix" object for controls set)
}
