% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalpp.R
\name{marginalpp.models}
\alias{marginalpp.models}
\title{Marginal PP for models sharing information between diseases}
\usage{
marginalpp.models(M, ABF, pr, kappa, p0)
}
\arguments{
\item{M}{list of model matrices for diseases 1, 2, ..., n}

\item{ABF}{list of log(ABF) vectors for diseases 1, 2, ...}

\item{pr}{list of prior probabilities for the models in M}

\item{kappa}{single value or vector of values to consider for the
sharing scale parameter}

\item{p0}{prior probability of the null model}
}
\value{
list of:
\itemize{
\item single.pp: list of pp for each model in \code{M[[i]]} for
disease i
\item shared.pp: list of pp for each model in \code{M[[i]]} for
disease i, M (not quite as input, reordered so null model is
first row
\item ABF: not quite as input, repordered so null model
is first
\item M: reordered so null model is first row
\item kappa: as supplied
}
}
\description{
Calculate marginal model posterior probabilities for each disease
}
\details{
Given a list of model matrices and log ABFs, this function
calculates the marginal model posterior probabilities for each
disease without ever calculating the joint Bayes Factors for all
cross-disease model configurations, which would require large
amounts of memory.
}
\author{
Chris Wallace
}
