% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalpp.R
\name{marginalpp}
\alias{marginalpp}
\title{Marginal PP for models sharing information between diseases}
\usage{
marginalpp(STR, ABF, pr, kappa, p0)
}
\arguments{
\item{STR}{list of models for diseases 1, 2, ..., n, each given in
the form of a character vector, with entries
\code{"snp1\%snp2\%snp3"}. The null model is given by \code{"0"}}

\item{ABF}{list of log(ABF) vectors for diseases 1, 2, ...}

\item{pr}{list of prior probabilities for the models in M}

\item{kappa}{single value or vector of values to consider for the
sharing scale parameter}

\item{p0}{prior probability of the null model}
}
\value{
list of:
single.pp: list of pp for each model in STR[[i]] for disease i
shared.pp: list of pp for each model in STR[[i]]
for disease i,
STR (not quite as input, reordered so null model
is first row
ABF: not quite as input, repordered so null
model is first row
kappa: as supplied

[[i]: R:[i
[[i]: R:[i
}
\description{
Calculate marginal model posterior probabilities for each disease
}
\details{
Given a list of model matrices and log ABFs, this function
calculates the marginal model posterior probabilities for each
disease without ever calculating the joint Bayes Factors for all
cross-disease model configurations, which would require large
amounts of memory.
}
\author{
Chris Wallace
}
