% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTFM-functions.R
\name{mod.split.fn}
\alias{mod.split.fn}
\title{when two traits, tnum is 1 or 2 , i.e. 1st or 2nd trait
called by PP.marg.fn}
\usage{
mod.split.fn(k, PP, tnum)
}
\description{
when two traits, tnum is 1 or 2 , i.e. 1st or 2nd trait
called by PP.marg.fn
}
