% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{MPP.fn}
\alias{MPP.fn}
\title{Find marginal PP of inclusion for SNPs (MPPi) in marginal models at each kappa value}
\usage{
MPP.fn(PP1)
}
\arguments{
\item{PP1}{is a matrix of marginal model PP for a trait (i.e. one list component of output from sharedmPP.fn),}
}
\value{
a matrix of MPPi, where rows are SNPs and columns are kappa
}
\description{
Find marginal PP of inclusion for SNPs (MPPi) in marginal models at each kappa value
}
\author{
Jenn Asimit
}
