% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{MPP.PP.groups.fn}
\alias{MPP.PP.groups.fn}
\title{Summarise, for a set of diseases, posterior probabilities and marginal posterior probabilities by SNP groups}
\usage{
MPP.PP.groups.fn(MPP, pp, dis, shared, snpGroups)
}
\arguments{
\item{MPP}{Matrix of marginal probabilities for a set of diseases and kappa values; output from MPP.fn}

\item{pp}{List consisting of posterior probability matrices for each disease; output from PP.fn}

\item{dis}{Vector of disease names}

\item{shared}{Vector of kappa values}
}
\value{
List consisting of mppGS: matrix of marginal probabilities by SNP groups; gPP: list of PP matrices by SNP groups
}
