% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa.R
\name{p.fn}
\alias{p.fn}
\title{Prior probability (in log units) for single trait models}
\usage{
p.fn(x, n, v)
}
\arguments{
\item{x}{Vector of number of SNPs in each model}

\item{n}{Number of SNPs in region}

\item{v}{Expected number of causal variants for region}
}
\value{
Vector of log prior probabilities for each single trait model of given size
}
\description{
Prior probability (in log units) for single trait models
}
