% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{pp.filter.fn}
\alias{pp.filter.fn}
\title{Filter probability matrix output by threshold}
\usage{
pp.filter.fn(pp, pthr)
}
\arguments{
\item{pp}{Matrix of probabilities: rows are kappa values; columns are SNPs (MPP matrix) or models (PP matrix)}

\item{pthr}{Threshold for keeping probabilities from SNPs/models that are greater than pthr for at least one kappa value}
}
\value{
Probability matrix with the same rows as input matrix, but only columns with the maximum probability greater than the threshold are retained.
}
\description{
Filter probability matrix output by threshold
}
