% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{PP.plot.fn}
\alias{PP.plot.fn}
\title{Plot a matrix of probabilities, where rows are kappa values and columns are SNPs (MPP matrix) or models (PP matrix)}
\usage{
PP.plot.fn(pp, pthr)
}
\arguments{
\item{pp}{Matrix of probabilities: rows are kappa values; columns are SNPs (MPP matrix) or models (PP matrix)}

\item{pthr}{Threshold for keeping probabilities from SNPs/models that are greater than pthr for at least one kappa value}
}
\value{
Matrix of probabilites that are plotted (only SNPs/models with MPP/PP > pthr at some kappa value are retained)
}
\description{
Plot a matrix of probabilities, where rows are kappa values and columns are SNPs (MPP matrix) or models (PP matrix)
}
\author{
Jenn Asimit
}
