% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTFM-functions.R
\name{PPBF.tags.fn}
\alias{PPBF.tags.fn}
\title{Calculate joint BFs and joint PP at tag SNP models}
\usage{
PPBF.tags.fn(t1snp.data, t1pheno, t2snp.data, t2pheno, tags, mppthr, mT1, MT1,
  mT2, MT2, trait1 = "T1", trait2 = "T2")
}
\arguments{
\item{t1snp.data}{a SnpMatrix object for trait 1}

\item{t1pheno}{a phenotype vector for trait 1}

\item{t2snp.data}{a SnpMatrix object for trait 2}

\item{t2pheno}{a phenotype vector for trait 2}

\item{tags}{tag SNPs from common controls}

\item{mppthr}{threshold for "best" SNPs MPP}

\item{mT1}{minimum model size for trait 1}

\item{MT1}{maximum model size for trait 1}

\item{mT2}{minimum model size for trait 2}

\item{MT2}{maximum model size for trait 2}
}
\value{
a list with components pp (data.frame of joint PP, prior,
logprior, logBF, and for each trait, model, logBF, model size)
and logBF0, the offset term needed to calculate joint BFs
}
\author{
Jenn Asimit
find "best" tag snps
generate all models from tag snps
}
