% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{PPmarginal.multiple.fast.fn}
\alias{PPmarginal.multiple.fast.fn}
\title{Approximate (fast) marginal PP for models of a set of diseases, sharing information between the diseases}
\usage{
PPmarginal.multiple.fast.fn(SM2, dis, thr, kappa, fthr, N0, ND, nsnps)
}
\arguments{
\item{SM2}{List of snpmod objects for a set of diseases}

\item{dis}{Vector of diseases for fine-mapping (subset from those in SM2)}

\item{thr}{Threshold such that the smallest set of models has cumulative PP >= thr}

\item{kappa}{Vector of sharing values}

\item{fthr}{Second level filtering on all but first disease to give a faster approximation; retain SNPs within the smallest set of models such that cumulative PP >= thr}

\item{N0}{number of shared controls}

\item{ND}{list of number of cases for a set of diseases}

\item{nsnps}{Number of SNPs in region}
}
\value{
List consisting of PP: marginal PP for models and MPP: marginal PP of SNP inclusion
}
\description{
Approximate (fast) marginal PP for models of a set of diseases, sharing information between the diseases
}
\author{
Jenn Asimit
}
