% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTFM-functions.R
\name{PPshared.fn}
\alias{PPshared.fn}
\title{Calculate joint PP at a set of sharing scales}
\usage{
PPshared.fn(shared, bf, tags, trait1, trait2, mydir)
}
\arguments{
\item{shared}{vector of sharing scales}

\item{bf}{output from BFexpanded.fn}

\item{tags}{set of tag SNPs for control set}

\item{trait1}{name of trait 1, for filename purposes}

\item{trait2}{name of trait 2, for filename purposes}

\item{mydir}{directory to save output to}
}
\value{
a list of PPmarg (PP for each trait) and mpp (marginal PP
for each trait) evalauted at each sharing scale value
}
\author{
Jenn Asimit
}
