% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTFM-functions.R
\name{sharedmPP.fn}
\alias{sharedmPP.fn}
\title{Find best tag SNP models, then expand and calculate joint PP at these expanded models}
\usage{
sharedmPP.fn(kappa, snp.data.list, pheno.list, tags, mppthr, traits, mydir)
}
\arguments{
\item{kappa}{is a vector of sharing scale values for the joint prior}

\item{snp.data.list}{is a list of SnpMatrix objects for each trait}

\item{pheno.list}{is a list of phenotype vectors for each trait}

\item{tags}{tags object of SNPs from common controls}

\item{mppthr}{threshold for "best" SNPs marginal PP of inclusion}

\item{traits}{is a vector of trait names}

\item{mydir}{is a directory stub to save marginal stochastic search (GUESSFM) results on tag SNPs}
}
\value{
a list with components pp1, pp2 (each is a data.frame of the marginal model PPs at each kappa value)
}
\author{
Jenn Asimit
find "best" tag snps
find min and max model sizes for each trait
generate all models from "best" tag snps
}
