% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SM2.R
\docType{data}
\name{SM2}
\alias{SM2}
\title{A list of GUESSFM output for an instance of simulated data}
\format{An object of class \code{list} of length 2.}
\usage{
SM2
}
\description{
Simulated data was generated using functions in MFMextra. In this example, there are two diseases that each have two
causal variants, of which one is shared beween diseases; both have causal variant rs61839660 in
group A; disease 1 has additional causal variant rs56382813 in group D; disease 2 has additional causal variant rs11594656 in
group C (SNP groups are given in the snpGroups object of this package). There are 3000 each of disease 1, disease 2,
and controls.
}
\details{
List items are "AD" and "AC", which are the GUESSFM output from each disease. Steps to obtain this are given in the vignettes.
}
\keyword{datasets}
