% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpGroups.R
\docType{data}
\name{snpGroups}
\alias{snpGroups}
\title{A list of 10 SNP groups in the IL2RA gene.}
\format{An object of class \code{list} of length 10.}
\usage{
snpGroups
}
\description{
Groups A-F were identified in the fine-mapping of multiple sclerosis (MS) and type 1
}
\keyword{datasets}
