% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTFM-functions.R
\name{T1mods.fn}
\alias{T1mods.fn}
\title{Generate all possible models for model sizes in a pre-specified range}
\usage{
T1mods.fn(mT1, MT1, msnps)
}
\arguments{
\item{mT1}{is the minimum model size}

\item{MT1}{is the maximum model size}

\item{msnps}{is the vector of SNPs that will be selected for models}
}
\value{
data.frame of models (e.g. snp1%snp2 for a 2-SNP model of snp1 and snp2) and their sizes
}
\author{
Jenn Asimit
if(mT1 > 0) T1mods <- NULL
if(mT1 == 0) T1mods <- matrix(rep(0,s),nrow=1)
}
