% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTFM-functions.R
\name{T1T2abfexp.fn}
\alias{T1T2abfexp.fn}
\title{use to combine expanded model bfs from traits; all models}
\usage{
T1T2abfexp.fn(T1bfexp, T2bfexp, logBF0)
}
\description{
use to combine expanded model bfs from traits; all models
}
