% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTFM-functions.R
\name{T1T2bfexp.best.fn}
\alias{T1T2bfexp.best.fn}
\title{on expanded marginal models, re-calculate joint bfs at "best" joint models so have joint bfs on expanded joint models}
\usage{
T1T2bfexp.best.fn(T1T2abf, T1bfexp, T2bfexp, indbest, logBF0)
}
\description{
on expanded marginal models, re-calculate joint bfs at "best" joint models so have joint bfs on expanded joint models
}
