% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTFM-functions.R
\name{T1T2bfexp.low.fn}
\alias{T1T2bfexp.low.fn}
\title{at expanded marginal models with low PP (keep same bf as tag model) use counts of no. of expanded models to get bfs (and counts of these models) at  expanded joint models}
\usage{
T1T2bfexp.low.fn(BFkeep, tags, logBF0)
}
\description{
at expanded marginal models with low PP (keep same bf as tag model) use counts of no. of expanded models to get bfs (and counts of these models) at  expanded joint models
}
