% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalpp.R
\name{calcpp}
\alias{calcpp}
\title{calcpp}
\usage{
calcpp(pr, lBF, norm = TRUE)
}
\arguments{
\item{pr}{vector or matrix of priors}

\item{lBF}{vector of log ABF}

\item{norm}{if TRUE (default) normalise so sum(pp) (or rowSum(pp)
if matrix) is 1. otherwise just return pr * exp(lBF)}
}
\value{
vector or matrix of pp
}
\description{
Internal function to combine vector or matrix prior with vector of log ABF to get pp per model
}
\author{
Chris Wallace
}
