% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTFM-functions.R
\name{gfm.sel.snps.fn}
\alias{gfm.sel.snps.fn}
\title{find tag SNPs with MPP > mppthr after running GUESSFM on tag SNPs and save to file
print table of posterior distribution for number of causal variants in GFM models
snpG is a "SnpMatrix" object; y is the phenotype vector; mppthr is the threshold for selecting
the "best" SNPs with marginal PP > mppthr}
\usage{
gfm.sel.snps.fn(snpG, y, tags, mppthr = 0.001)
}
\description{
find tag SNPs with MPP > mppthr after running GUESSFM on tag SNPs and save to file
print table of posterior distribution for number of causal variants in GFM models
snpG is a "SnpMatrix" object; y is the phenotype vector; mppthr is the threshold for selecting
the "best" SNPs with marginal PP > mppthr
}
