% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginalpp.R
\name{marginalpp}
\alias{marginalpp}
\title{Marginal PP for models sharing information between diseases}
\usage{
marginalpp(STR, ABF, pr, kappa, p0, N0, ND, nsnps, I0 = as.list(rep(0,
  length(ND))))
}
\arguments{
\item{STR}{list of models for diseases 1, 2, ..., n, each given in
the form of a character vector, with entries
\code{"snp1\%snp2\%snp3"}. The null model is given by
\code{"1"} OR \code{"0"}.  It is assumed that all elements of
ABF, PP and pr below follow this same order.}

\item{ABF}{list of log(ABF) vectors for diseases 1, 2, ...}

\item{pr}{list of prior probabilities for the models in M}

\item{kappa}{single value or vector of values to consider for the
sharing scale parameter.  the value of kappa=1 must be
included, and if not will be prepended.}

\item{p0}{prior probability of the null model}

\item{N0}{number of shared controls}

\item{ND}{list of number of cases for a set of diseases}

\item{nsnps}{number of snps in region}

\item{I0}{list of number of controls for a set of diseases that are only used as controls for a specific disease (ie not shared)
...}
}
\value{
list of: - single.pp: list of pp for each model in
\code{STR[[i]]} for disease i - shared.pp: list of pp for each model
in \code{STR[[i]]} for disease i, - STR: not quite as input,
reordered so null model is first row - ABF: not quite as
input, repordered so null model is first row - kappa: as
supplied
}
\description{
Calculate marginal model posterior probabilities for each disease
}
\details{
Given a list of model matrices and log ABFs, this function
calculates the marginal model posterior probabilities for each
disease without ever calculating the joint Bayes Factors for all
cross-disease model configurations, which would require large
amounts of memory.
}
\author{
Chris Wallace
}
