% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{PPmarginal.multiple.fn}
\alias{PPmarginal.multiple.fn}
\title{Marginal PP for models of a set of diseases, sharing information between the diseases}
\usage{
PPmarginal.multiple.fn(SM2, dis, thr, TOdds, N0, ND, nsnps)
}
\arguments{
\item{SM2}{List of snpmod objects for a set of diseases}

\item{dis}{Vector of diseases for fine-mapping (subset from those in SM2)}

\item{thr}{Threshold such that the smallest set of models has cumulative PP >= thr}

\item{TOdds}{Vector of target odds of no sharing to sharing
...}

\item{N0}{number of shared controls}

\item{ND}{list of number of cases for a set of diseases}

\item{nsnps}{number of SNPs in the region}
}
\value{
List consisting of PP: marginal PP for models and MPP: marginal PP of SNP inclusion
}
\description{
Marginal PP for models of a set of diseases, sharing information between the diseases
}
\author{
Jenn Asimit
}
