% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTFM-functions.R
\name{T1mods.fn}
\alias{T1mods.fn}
\title{mT1 and MT1 are the min and max number of causals in a model for trait t1
msnps is the vector of snps to consider in the models}
\usage{
T1mods.fn(mT1, MT1, msnps)
}
\description{
mT1 and MT1 are the min and max number of causals in a model for trait t1
msnps is the vector of snps to consider in the models
}
